Split comics definition and user databases
Users are now stored separately from comics databases, and the default
load/dump commands only load/dump the comics database. This is because
separate instances of the system will share and want to be able to merge
their comics databases, but the user data is very much separate and
system-specific.